refactor: SocialLoginResult 패키징 변경
