+++ src/main/java/com/mobble/mobbleserver/application/account/provided/SocialLoginPort.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.application.account.provided;
 
-import com.mobble.mobbleserver.application.account.command.SocialLoginResult;
+import com.mobble.mobbleserver.application.account.result.SocialLoginResult;
 import com.mobble.mobbleserver.infrastructure.web.account.dto.request.SocialLoginRequestDto;
 
 public interface SocialLoginPort {
