+++ src/main/java/com/mobble/mobbleserver/infrastructure/web/account/SocialLoginAPI.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.infrastructure.web.account;
 
-import com.mobble.mobbleserver.application.account.command.SocialLoginResult;
+import com.mobble.mobbleserver.application.account.result.SocialLoginResult;
 import com.mobble.mobbleserver.application.account.provided.SocialLoginPort;
 import com.mobble.mobbleserver.infrastructure.web.account.dto.request.SocialLoginRequestDto;
 import jakarta.validation.Valid;
